requirements.txt
openai
twilio
fastapi
uvicorn
requests
gspread
oauth2client
python-dotenv
python-multipart
boto3
openai==0.28
